Separate TypeDeclFinder from AccessScopeChecker (#23957)
...and move the latter into lib/Sema as TypeAccessScopeChecker, since
it's not interesting outside of Sema anyway.

No functionality change.
